+++ Akalat-Shop/Model/Driver/Responses/ReadyOrders/DriverMealDetails.swift
@@ -12,11 +12,12 @@ For support, please feel free to contact me at https://www.linkedin.com/in/syeda
 */
 
 import Foundation
+
 struct DriverMealDetails : Codable {
     
 	let id               : Int
 	let name             : String
-	let price            : Int
+	let price            : Double
 
 	enum CodingKeys: String, CodingKey {
